refactor(core): split up server and app, fix year
