+++ src/views/index.ejs
@@ -178,7 +178,7 @@
         </div>
         
         <footer class="footer text-center">
-            <p>MultiZlogin &copy; 2023 - Hệ thống quản lý đa tài khoản Zalo</p>
+            <p>MultiZlogin &copy; 2025 - Hệ thống quản lý đa tài khoản Zalo</p>
         </footer>
     </div>
